refactor: NotificationToken DeviceType enum + PredictRequest BigDecimal 정합

+++ backend/src/main/java/com/ssafy/s309/domain/prediction/client/dto/FoodNutrition.java
@@ -1,13 +1,11 @@
 package com.ssafy.s309.domain.prediction.client.dto;
 
-import java.math.BigDecimal;
-
 public record FoodNutrition(
     String foodId,
     String name,
-    BigDecimal carbsG,
-    BigDecimal proteinG,
-    BigDecimal fatG,
-    BigDecimal kcal,
-    BigDecimal sugarG,
+    Double carbsG,
+    Double proteinG,
+    Double fatG,
+    Double kcal,
+    Double sugarG,
     Integer giScore) {}
